+++ Src/Library/changelog.md
@@ -1,5 +1,6 @@
 ### NEW
 - add `ProducesProblemFE()` extension method for `RouteHandlerBuilder` [#info](https://discord.com/channels/933662816458645504/1004762111546769498)
+- bult-in unhandled exception handler now sends a response of type `InternalErrorResponse`
 
 ### IMPROVEMENTS
 - increase logging in validation schema processor #117